requirements.txt
requests==2.26.0
GitPython==3.1.18
prometheus-client==0.11.0
rich==10.9.0
schedule==1.1.0
